Corrected bugs and errors introduced by previous commits
Please, watch out your commits.

The unix convention to say that something is an argument is to put it in
upper case: target COUNTRY
Putting it between brackets means that it is optional, which it was not.

Furthermore, your commit didn't pass the unit test, be sure to run it
systematically !