Added the very beginning of cli_ui
cli_ui intends to be a cli interface to the game where you have no turn
(those are computed by looking at the time between two commands) and the game
is saved at each command (all of that to maintain the illusion of continuity of
the game).

Only the help is present yet but with docopt it provides good guidelines to
implement the user interface.