Added a server to be used in multiplayer versions of the game
WARNING!
This server has not been tested yet, it lacks a multiplayer-friendly
implementation of World.

The guide-line here is to implement MultiWorld as an alternative world, test
it, and then modify the whole programm to use only MultiWorld and not World
so that we have only one code to support.